local_data_processing/requirements.txt
pybaseball>=2.2.7
google-cloud-bigquery>=3.11.0
pandas>=2.0.0
numpy>=1.24.0
